feat: AssertionScope test refactoring from PR #79 feedback #186047
